+++ src/installExt.js
@@ -6,6 +6,7 @@ const childProcess = require('child_process');
 const path = require('path');
 const fs = require('fs');
 const { typeExtMapper, codeExt } = require('./config');
+const myLog = require('./myLog');
 
 const extFileNames = fs.readdirSync(path.resolve(__dirname, './vscodeExts'));
 
@@ -17,17 +18,17 @@ installedExts = installedExts.toString().split('\n');
  * @param {string} extName
  */
 function installVscodeExtension(extName) {
-  console.log(`Installing vscode plugin ${extName}...`);
+  myLog.start(`Installing vscode plugin ${extName}...`);
   try {
     // find extension .vsix file
     const vsixName = extFileNames.find(str => str.startsWith(extName));
     // if exist .vsix file ,use it. if not, auto download in vscode extension market
     let cmd = 'code --install-extension ' + (vsixName ? path.resolve(__dirname, './vscodeExts/', vsixName) : extName);
     childProcess.execSync(cmd);
-    console.log(`✔ Vscode plugin(${extName}) installation succeeded`);
+    myLog.success(`Vscode plugin(${extName}) installation succeeded`);
   } catch (err) {
     // maybe network error
-    console.log('✘ Install vscode plugin failed: ' + err);
+    myLog.danger('Install vscode plugin failed: ' + err);
   }
 }
 /**
@@ -44,10 +45,10 @@ function checkExtension(type) {
       const installedExtVersion = installedExt.split('@')[1];
       // check installed extension's version
       if (installedExtVersion < codeExtItem.validVersion) {
-        console.warn(`Installed plugin version is too early(${installedExt})`);
+        myLog.warn(`Installed plugin version is too early(${installedExt})`);
         installVscodeExtension(codeExtItem.name);
       } else {
-        console.log('✔ Installed vscode plugin:', installedExt, '. Auto skip this stage');
+        myLog.success('Installed vscode plugin:', installedExt, '. Auto skip this stage');
       }
     } else {
       installVscodeExtension(codeExtItem.name);
@@ -62,11 +63,6 @@ module.exports = function ({ type }) {
   checkExtension('eslint');
   checkExtension(type);
   if (type.startsWith('vue')) {
-    let clgMap = {
-      vue2: '■ Please disable volar and enable vetur.',
-      vue3: '■ Please disable vetur and enable volar.',
-    };
-    console.log(clgMap[type.substring(0, 4)]);
-    console.log('■ If eslint not work with vue-cli, Try to remove @vue/cli-plugin-eslint');
+    myLog.tip('If eslint not work with vue-cli, Try to remove @vue/cli-plugin-eslint');
   }
 };