+++ src/installNpmPkgs.js
@@ -6,6 +6,8 @@ const childProcess = require('child_process');
 const { npmPkgs } = require('./config');
 const chalk = require('chalk');
 const inquirer = require('inquirer');
+const { readFileSync } = require('fs');
+const myLog = require('./myLog');
 const prompt = inquirer.createPromptModule();
 
 module.exports = async function ({ type }) {
@@ -19,7 +21,7 @@ module.exports = async function ({ type }) {
   pkgTypes.push(type);
 
   if (await checkTypescriptVersion(type)) {
-    pkgTypes.push('tsBase');
+    pkgTypes.push('tsBase'); // 基本ts依赖
   }
 
   for (const key of pkgTypes) {
@@ -32,31 +34,39 @@ module.exports = async function ({ type }) {
   console.log(npmCmd);
   try {
     childProcess.execSync(npmCmd);
-    console.log('✔ Npm packages installation succeed');
+    myLog.success('Npm packages installation succeed');
   } catch (err) {
-    if (err) return console.error('✘ Npm install failed', err);
+    if (err) return myLog.danger('Npm install failed', err);
   }
 };
 
 /** 提示用户是否安装typescript */
 async function checkTypescriptVersion(type) {
   if (type.indexOf('ts') > -1) {
-    console.log('Checking typescript...');
+    myLog.start('Checking typescript...');
     // check user if install the typescript dep
-    const cmdReturn = childProcess.execSync('npm list typescript').toString();
-    const tsInfo = cmdReturn.match(/typescript@\d(\.\d{1,2}){2}/);
-    let defaultInstallTs = false;
-    if (tsInfo) {
-      console.log('Currently installed: ', chalk.green(tsInfo[0].toLowerCase()));
+    let tsVersion;
+    try {
+      let tsPackageStr = readFileSync('./node_modules/typescript/package.json');
+      let tsPackageJson = JSON.parse(tsPackageStr);
+      tsVersion = tsPackageJson.version;
+      // const cmdReturn = childProcess.execSync('npm list typescript').toString();
+    } catch (e) {
+      // -2 not found file or directory
+      if (e.errno != -2) {
+        myLog.danger('Can not get typescript info');
+      }
+    }
+    if (tsVersion) {
+      myLog.success('Currently installed: typescript@', chalk.green(tsVersion));
     } else {
-      defaultInstallTs = true;
-      console.log('Typescript not currently installed.');
+      myLog.warn('Typescript not currently installed.');
     }
     const answer = await prompt({
       type: 'confirm',
       name: 'installTs',
       message: `Install typescript@${npmPkgs.tsBase.typescript}?`,
-      default: defaultInstallTs,
+      default: !tsVersion,
     });
     return answer.installTs;
   }